chore: use central caching for everything
