+++ crates/ql_core/Cargo.toml
@@ -28,7 +28,6 @@ serde.workspace = true
 serde_json.workspace = true
 thiserror.workspace = true
 semver.workspace = true
-sha2 = "0.10"
 http-cache-reqwest = { version = "1.0.0-alpha.6" }
 reqwest-middleware = "0.5.1"
