+++ crates/ql_core/src/lib.rs
@@ -46,7 +46,7 @@ mod progress;
 pub mod read_log;
 pub mod request;
 mod structs;
-pub mod urlcache;
+pub mod url;
 
 pub use crate::json::InstanceConfigJson;
 pub use constants::*;
